Move the Viewport a ViewBox classes in the org.girod.javafx.svgimage package Fix #36: Allow to scale on parent element
