+++ src/core/org/girod/javafx/svgimage/xml/specs/MarkerSpec.java
@@ -32,6 +32,8 @@ the project website at the project page on https://github.com/hervegirod/fxsvgim
  */
 package org.girod.javafx.svgimage.xml.specs;
 
+import org.girod.javafx.svgimage.Viewbox;
+import org.girod.javafx.svgimage.Viewport;
 import org.girod.javafx.svgimage.xml.parsers.ParserUtils;
 import org.girod.javafx.svgimage.xml.parsers.SVGTags;
 import org.girod.javafx.svgimage.xml.parsers.XMLNode;
