+++ web/source/config.properties
@@ -1,0 +1,1 @@
+javadoc-api=javadoc
