+++ web/source/history.xml
@@ -179,11 +179,16 @@
       
    <title title="Version 1.2" /> 
    <ul>  
+      <li>Use docJGenerator 1.6.10</li>
+      <li>Use groovy 4.0.26 (only uses for tests, the binaries have no dependencies</li>
       <li>Fix #47: Add a parameter for the Color of the svg background in the JavaFX to SVG parameters</li> 
       <li>Fix #48: Add a parameter to generate the viewbox in the JavaFX to SVG parameters</li>   
       <li>Fix #49: Add a parameter to not allow the transform on the root of the svg document in the JavaFX to SVG parameters</li>   
-      <li>Fix #42: Failed to render text with tspan</li>
-      <li>Fix #45: Mask not working</li>                
+      <li>Fix #42: Failed to render text with tspan in some cases</li>
+      <li>Fix #45: Mask not working</li>     
+      <li>Fix #37: SVGImage's scaleTo doesn't seem to override svg width height parameters</li>   
+      <li>Fix #50: The default LineCap, LineJoin, and miterlimit on Shapes are incorrect</li> 
+      <li>Fix #36: Allow to scale on parent element</li>   
    </ul>        
    
    <cat id="general" />
